Added synchronization to checklist export and import
